Tide gauges for tsunami example
Ignore-this: ab19c896243aaaeb611e9de31da902f2

darcs-hash:0b46c9c1280bf4f9ea394e25155bfd1afb00a2d8

+++ utils.h
@@ -208,7 +208,7 @@ timer timer_start (void)
   return t;
 }
 
-void timer_print (timer t, int i, int tnc)
+void timer_print (timer t, int i, long tnc)
 {
   clock_t end = clock ();
   struct timeval tvend;
